+++ http_lazy_headers/fields/date.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from ..shared.common import dates
 from ..shared import bases
 
 
-class Date(bases.SingleHeaderBase):
+class Date(bases.DateSomeBase):
     """
     The ``Date`` header field represents\
     the date and time at which the message was originated.
@@ -19,12 +18,3 @@ class Date(bases.SingleHeaderBase):
     """
 
     name = 'date'
-
-    def check_value(self, value):
-        dates.check_date(value)
-
-    def values_str(self, values):
-        return dates.format_date(values[0])
-
-    def clean_value(self, raw_value):
-        return dates.clean_date_time(raw_value)
